requirements.txt
django
djangorestframework
psycopg2-binary
openai
pillow
python-decouple
dj-database-url
gunicorn
